Move state logic to backend
Frontend only sends the position of the click to the backend.
The backend will save the location of a previous click if the previous position was on a valid figure. With this
information, it is clear what figure should go where. The frontend then only needs to draw the objects and is completely
stateless.
Also removes the need for a figure id